+++ src/virtual_dom/update_el.rs
@@ -1,4 +1,4 @@
-use super::{Attrs, El, ElRef, EventHandler, Node, Style, Tag, Text};
+use super::{Attrs, El, ElKey, ElRef, EventHandler, Node, Style, Tag, Text};
 
 /// `UpdateEl` is used to distinguish arguments in element-creation macros, and handle
 /// each type appropriately.
@@ -130,6 +130,12 @@ impl<Ms> UpdateEl<El<Ms>> for Tag {
 impl<Ms, E: Clone> UpdateEl<El<Ms>> for ElRef<E> {
     fn update(self, el: &mut El<Ms>) {
         el.refs.push(self.shared_node_ws);
+    }
+}
+
+impl<Ms> UpdateEl<El<Ms>> for ElKey {
+    fn update(self, el: &mut El<Ms>) {
+        el.key = Some(self);
     }
 }